backend/requirements.txt
fastapi==0.128.1
uvicorn==0.40.0
sqlmodel==0.0.32
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.2.1
psycopg2-binary
